Refactored timezone functions.
