+++ libfreerdp/core/settings.c
@@ -472,7 +472,7 @@ rdpSettings* freerdp_settings_new(DWORD flags)
 	if (!settings->ServerAutoReconnectCookie)
 			goto out_fail;
 
-	settings->ClientTimeZone = (TIME_ZONE_INFO*) calloc(1,sizeof(TIME_ZONE_INFO));
+	settings->ClientTimeZone = (LPTIME_ZONE_INFORMATION) calloc(1,sizeof(TIME_ZONE_INFORMATION));
 	if (!settings->ClientTimeZone)
 			goto out_fail;
 
@@ -762,10 +762,10 @@ rdpSettings* freerdp_settings_clone(rdpSettings* settings)
 		CopyMemory(_settings->ClientAutoReconnectCookie, settings->ClientAutoReconnectCookie, sizeof(ARC_CS_PRIVATE_PACKET));
 		CopyMemory(_settings->ServerAutoReconnectCookie, settings->ServerAutoReconnectCookie, sizeof(ARC_SC_PRIVATE_PACKET));
 
-		_settings->ClientTimeZone = (TIME_ZONE_INFO*) malloc(sizeof(TIME_ZONE_INFO));
+		_settings->ClientTimeZone = (LPTIME_ZONE_INFORMATION) malloc(sizeof(TIME_ZONE_INFORMATION));
 		if (!_settings->ClientTimeZone)
 			goto out_fail;
-		CopyMemory(_settings->ClientTimeZone, settings->ClientTimeZone, sizeof(TIME_ZONE_INFO));
+		CopyMemory(_settings->ClientTimeZone, settings->ClientTimeZone, sizeof(TIME_ZONE_INFORMATION));
 
 		_settings->TargetNetAddressCount = settings->TargetNetAddressCount;
 
@@ -815,7 +815,7 @@ rdpSettings* freerdp_settings_clone(rdpSettings* settings)
 			goto out_fail;
 		}
 
- 		if (_settings->DeviceArraySize < _settings->DeviceCount)
+		if (_settings->DeviceArraySize < _settings->DeviceCount)
 		{
 			_settings->DeviceCount = 0;
 			_settings->DeviceArraySize = 0;
