+++ libfreerdp/core/timezone.c
@@ -22,8 +22,12 @@
 #endif
 
 #include <winpr/crt.h>
+#include <winpr/timezone.h>
 
 #include "timezone.h"
+
+static void rdp_read_system_time(wStream* s, SYSTEMTIME* system_time);
+static void rdp_write_system_time(wStream* s, SYSTEMTIME* system_time);
 
 /**
  * Read SYSTEM_TIME structure (TS_SYSTEMTIME).\n
@@ -32,7 +36,7 @@
  * @param system_time system time structure
  */
 
-void rdp_read_system_time(wStream* s, SYSTEM_TIME* system_time)
+void rdp_read_system_time(wStream* s, SYSTEMTIME* system_time)
 {
 	Stream_Read_UINT16(s, system_time->wYear); /* wYear, must be set to 0 */
 	Stream_Read_UINT16(s, system_time->wMonth); /* wMonth */
@@ -51,7 +55,7 @@ void rdp_read_system_time(wStream* s, SYSTEM_TIME* system_time)
  * @param system_time system time structure
  */
 
-void rdp_write_system_time(wStream* s, SYSTEM_TIME* system_time)
+void rdp_write_system_time(wStream* s, SYSTEMTIME* system_time)
 {
 	Stream_Write_UINT16(s, system_time->wYear); /* wYear, must be set to 0 */
 	Stream_Write_UINT16(s, system_time->wMonth); /* wMonth */
@@ -76,34 +80,26 @@ void rdp_write_system_time(wStream* s, SYSTEM_TIME* system_time)
 
 BOOL rdp_read_client_time_zone(wStream* s, rdpSettings* settings)
 {
-	char* str = NULL;
-	TIME_ZONE_INFO* clientTimeZone;
+	LPTIME_ZONE_INFORMATION tz;
 
 	if (Stream_GetRemainingLength(s) < 172)
 		return FALSE;
 
-	clientTimeZone = settings->ClientTimeZone;
+	tz = settings->ClientTimeZone;
 
-	Stream_Read_UINT32(s, clientTimeZone->bias); /* Bias */
+	Stream_Read_UINT32(s, tz->Bias); /* Bias */
 
 	/* standardName (64 bytes) */
-	ConvertFromUnicode(CP_UTF8, 0, (WCHAR*) Stream_Pointer(s), 64 / 2, &str, 0, NULL, NULL);
-	Stream_Seek(s, 64);
-	strncpy(clientTimeZone->standardName, str, sizeof(clientTimeZone->standardName));
-	free(str);
-	str = NULL;
+	Stream_Read(s, tz->StandardName, sizeof(tz->StandardName));
 
-	rdp_read_system_time(s, &clientTimeZone->standardDate); /* StandardDate */
-	Stream_Read_UINT32(s, clientTimeZone->standardBias); /* StandardBias */
+	rdp_read_system_time(s, &tz->StandardDate); /* StandardDate */
+	Stream_Read_UINT32(s, tz->StandardBias); /* StandardBias */
 
 	/* daylightName (64 bytes) */
-	ConvertFromUnicode(CP_UTF8, 0, (WCHAR*) Stream_Pointer(s), 64 / 2, &str, 0, NULL, NULL);
-	Stream_Seek(s, 64);
-	strncpy(clientTimeZone->daylightName, str, sizeof(clientTimeZone->daylightName));
-	free(str);
+	Stream_Read(s, tz->DaylightName, sizeof(tz->DaylightName));
 
-	rdp_read_system_time(s, &clientTimeZone->daylightDate); /* DaylightDate */
-	Stream_Read_UINT32(s, clientTimeZone->daylightBias); /* DaylightBias */
+	rdp_read_system_time(s, &tz->DaylightDate); /* DaylightDate */
+	Stream_Read_UINT32(s, tz->DaylightBias); /* DaylightBias */
 
 	return TRUE;
 }
@@ -117,53 +113,37 @@ BOOL rdp_read_client_time_zone(wStream* s, rdpSettings* settings)
 
 void rdp_write_client_time_zone(wStream* s, rdpSettings* settings)
 {
-	WCHAR* standardName = NULL;
-	WCHAR* daylightName = NULL;
-	int standardNameLength;
-	int daylightNameLength;
-	TIME_ZONE_INFO* clientTimeZone;
+	LPTIME_ZONE_INFORMATION tz;
+	DWORD rc;
 
-	clientTimeZone = settings->ClientTimeZone;
-	freerdp_time_zone_detect(clientTimeZone);
+	tz = settings->ClientTimeZone;
+	rc = GetTimeZoneInformation(tz);
 
-	standardNameLength = ConvertToUnicode(CP_UTF8, 0, clientTimeZone->standardName, -1, &standardName, 0) * 2;
-	daylightNameLength = ConvertToUnicode(CP_UTF8, 0, clientTimeZone->daylightName, -1, &daylightName, 0) * 2;
-
-	if (standardNameLength > 62)
-		standardNameLength = 62;
-
-	if (daylightNameLength > 62)
-		daylightNameLength = 62;
-
-    /* Bias */
- 	Stream_Write_UINT32(s, clientTimeZone->bias);
+	/* Bias */
+	Stream_Write_UINT32(s, tz->Bias);
 
 	/* standardName (64 bytes) */
-	Stream_Write(s, standardName, standardNameLength);
-	Stream_Zero(s, 64 - standardNameLength);
+	Stream_Write(s, tz->StandardName, sizeof(tz->StandardName));
 
-    /* StandardDate */
-    rdp_write_system_time(s, &clientTimeZone->standardDate);
+	/* StandardDate */
+	rdp_write_system_time(s, &tz->StandardDate);
 
-	DEBUG_TIMEZONE("bias=%d stdName='%s' dlName='%s'", clientTimeZone->bias, clientTimeZone->standardName, clientTimeZone->daylightName);
+	DEBUG_TIMEZONE("bias=%d stdName='%s' dlName='%s'", tz->Bias,
+		       tz->StandardName, tz->DaylightName);
 
 	/* Note that StandardBias is ignored if no valid standardDate is provided. */
-    /* StandardBias */
-	Stream_Write_UINT32(s, clientTimeZone->standardBias);
-	DEBUG_TIMEZONE("StandardBias=%d", clientTimeZone->standardBias);
+	/* StandardBias */
+	Stream_Write_UINT32(s, tz->StandardBias);
+	DEBUG_TIMEZONE("StandardBias=%d", tz->StandardBias);
 
 	/* daylightName (64 bytes) */
-    Stream_Write(s, daylightName, daylightNameLength);
-    Stream_Zero(s, 64 - daylightNameLength);
+	Stream_Write(s, tz->DaylightName, sizeof(tz->DaylightName));
 
-     /* DaylightDate */
-    rdp_write_system_time(s, &clientTimeZone->daylightDate);
+	/* DaylightDate */
+	rdp_write_system_time(s, &tz->DaylightDate);
 
 	/* Note that DaylightBias is ignored if no valid daylightDate is provided. */
-    /* DaylightBias */
-	Stream_Write_UINT32(s, clientTimeZone->daylightBias);
-	DEBUG_TIMEZONE("DaylightBias=%d", clientTimeZone->daylightBias);
-
-	free(standardName);
-	free(daylightName);
+	/* DaylightBias */
+	Stream_Write_UINT32(s, tz->DaylightBias);
+	DEBUG_TIMEZONE("DaylightBias=%d", tz->DaylightBias);
 }
