+++ libfreerdp/core/timezone.h
@@ -28,13 +28,9 @@
 
 #include <freerdp/log.h>
 #include <freerdp/freerdp.h>
-#include <freerdp/locale/timezone.h>
 
 #include <winpr/stream.h>
 
-void rdp_read_system_time(wStream* s, SYSTEM_TIME* system_time);
-void rdp_write_system_time(wStream* s, SYSTEM_TIME* system_time);
-void rdp_get_client_time_zone(wStream* s, rdpSettings* settings);
 BOOL rdp_read_client_time_zone(wStream* s, rdpSettings* settings);
 void rdp_write_client_time_zone(wStream* s, rdpSettings* settings);
