+++ libfreerdp/locale/CMakeLists.txt
@@ -22,7 +22,6 @@ set(${MODULE_PREFIX}_SRCS
 	keyboard_layout.c
 	keyboard.c
 	locale.c
-	timezone.c
 	liblocale.h)
 
 set(${MODULE_PREFIX}_X11_SRCS
